Add task to delete related ES docs
When a Collection or DIP is deleted, check descendants existence and
delete the descendant documents in the related ES indexes from a Celery
task. Use low level client `delete_by_query` to be able to delete over
multiple indexes at the same time. For example, when a Collection is
deleted, to be able to delete the related DIPs and DigitalFiles, which
share the same field in their indexes for the Collection id.

- Add `delete_es_descendants` Celery task.
- Add abstract method to ES models to check need of descendants update.
- Call async. task when descendant update is needed.
- Show alert to user about delete in progress.
- Fix, improve and add tests.